refactor: put tahoe-relevant code in tahoe.rs
