+++ magic_cap/src/catalog.rs
@@ -1,5 +1,6 @@
 use crate::err::MagicCapError;
-use crate::{Immutable, ImmutableBuilder, ImmutableReadCap, ImmutableVerifyCap, tagged_hash};
+use crate::{Immutable, ImmutableBuilder, ImmutableReadCap, ImmutableVerifyCap};
+use crate::tahoe::tagged_hash;
 use data_encoding::HEXLOWER;
 use std::fmt;
 use std::fs::File;
